Moved the WriterFactory and its default implementation to the org.quickperf.writer package. Updated the imports accordingly. Put the default format in a constant in the annotation. Fixed the construction of the instance of the MeasureHeapAllocation annotation in the JvmAnnotationBuilder.
