+++ core/src/main/java/org/quickperf/writer/WriterFactory.java
@@ -1,4 +1,4 @@
-package org.quickperf.jvm.allocation;
+package org.quickperf.writer;
 
 import java.io.IOException;
 import java.io.Writer;
